separate every piece in a separate class, implemented game draw and fixed king tile bug
